Refactor code to reduce code duplication of ModelStub
