+++ src/main/java/seedu/address/logic/parser/EventCommandParser.java
@@ -6,10 +6,12 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_DATE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 
 import java.time.LocalDate;
+import java.time.format.DateTimeParseException;
 import java.util.Set;
 import java.util.stream.Stream;
 
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.Messages;
 import seedu.address.logic.commands.EventCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
@@ -27,19 +29,25 @@ public class EventCommandParser implements Parser<EventCommand> {
         ArgumentMultimap argMultimap =
                 ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_DATE, PREFIX_ATTENDEES);
 
-        if (!arePrefixesPresent(argMultimap, PREFIX_NAME, PREFIX_DATE, PREFIX_ATTENDEES)
+        if (!arePrefixesPresent(argMultimap, PREFIX_NAME, PREFIX_DATE)
                 || !argMultimap.getPreamble().isEmpty()) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, EventCommand.MESSAGE_USAGE));
         }
 
-        argMultimap.verifyNoDuplicatePrefixesFor(PREFIX_NAME, PREFIX_DATE, PREFIX_ATTENDEES);
+        argMultimap.verifyNoDuplicatePrefixesFor(PREFIX_NAME, PREFIX_DATE);
         String name = argMultimap.getValue(PREFIX_NAME).get().trim();
-        LocalDate date = LocalDate.parse(argMultimap.getValue(PREFIX_DATE).get().trim());
+        try {
+            LocalDate date = LocalDate.parse(argMultimap.getValue(PREFIX_DATE).get().trim());
 
-        String indexes = argMultimap.getValue(PREFIX_ATTENDEES).get();
-        Set<Index> attendeeIndexes = ParserUtil.parseIndexes(indexes);
+            String indexes = (arePrefixesPresent(argMultimap, PREFIX_ATTENDEES))
+                             ? argMultimap.getValue(PREFIX_ATTENDEES).get()
+                             : " ";
+            Set<Index> attendeeIndexes = ParserUtil.parseIndexes(indexes);
 
-        return new EventCommand(name, date, attendeeIndexes);
+            return new EventCommand(name, date, attendeeIndexes);
+        } catch (DateTimeParseException e) {
+            throw new ParseException(Messages.MESSAGE_INVALID_DATE_FORMAT);
+        }
     }
 
     /**
